+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `MessageOriginalTranslationsStore`: In-memory storage for keeping track of all auto-translated messages which have been manually converted to their original text. [#5842](https://github.com/GetStream/stream-chat-android/pull/5842)
 
 ### ⚠️ Changed
 
@@ -81,6 +82,7 @@
 - Introduce internal `DraftChannelViewController` and experimental classes `DraftChannelViewState`, `DraftChannelViewAction`, and `DraftChannelViewEvent`. [#5797](https://github.com/GetStream/stream-chat-android/pull/5797)
 - Message member from the member bottom sheet of the channel info screen. [#5797](https://github.com/GetStream/stream-chat-android/pull/5797)
 - Introduce **experimental** factory methods for the channel info screens. [#5812](https://github.com/GetStream/stream-chat-android/pull/5812)
+- Add `ChatTheme.showOriginalTranslationEnabled` flag for enabling/disabling the "Show original" translation label in the message list. [#5842](https://github.com/GetStream/stream-chat-android/pull/5842)
 
 ### ⚠️ Changed